Add visibility to config

+++ scripts/build.sh
@@ -10,9 +10,13 @@ fi
 prev=$(grep -o 'v\([0-9]\+.\)\{2\}\([0-9]\+\)' server/config/config.json | head -1)
 
 if [[ $# -gt 0 ]]; then
-	v=v$1
-  sed -i -e "s/$prev/$v/g" server/config/config.json
-  rm server/config/config.json-e
+  if [[ $1 -eq "auto"]]; then
+     rep=yes
+  else
+  	v=v$1
+    sed -i -e "s/$prev/$v/g" server/config/config.json
+    rm server/config/config.json-e
+  fi
 else
   v=$prev
 fi
@@ -23,10 +27,12 @@ ng build --prod
 echo 'Build Docker image'
 # Build docker
 docker build -t ibmcase/casewebportal .
-read -p 'Push to dockerhub?: (yes)/no' rep
+if [  -z "$rep" ]; then
+  read -p 'Push to dockerhub?: (yes)/no' rep
+fi
 if [ -z "$rep" ]
 then
-      echo "Pushing..."
-      docker login
-      docker push ibmcase/casewebportal:latest
+  echo "Pushing..."
+  docker login
+  docker push ibmcase/casewebportal:latest
 fi
